xp-backend/requirements.txt
fastapi
uvicorn[standard]
aiogram
python-dotenv